+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * BREAKING: rename `abstraction` to `protocol`
 * Move parsing operations to `PacketParser`
+* Allow reconnecting to an previously authorized `SerialPort`/`BluetoothDevice`
 
 # 0.0.1-alpha.45